requirements.txt
jira~=3.5.2
DateTime~=5.3
python-dotenv~=1.0.0
